requirements.txt
openai
requests
google-api-python-client
google-auth
resend
